pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calibract"
version = "0.1.0"
description = "Calibration-focused parameter-efficient transfer learning for CNNs with learnable activations."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Anton Nemchinski" }]
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "Pillow>=10.0",
    "openpyxl>=3.1",
    "matplotlib>=3.8",
    "PyYAML>=6.0",
    "torchsummary>=1.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]
